Don't use parallel shutdown in atexit function
Using parallel futures in an atexit callback is not supported by Python
and will result in an error like this being shown:
```
RuntimeError: cannot schedule new futures after interpreter shutdown
```

This was already fixed for arbitrary config tests by not shutting down
in parallel during atexit, but the same was not done for upgrade tests.
This uses the same approach in upgrade tests.